Refactor AppWebTestCase #39
